+++ pyanalyze/safe.py
@@ -8,7 +8,11 @@ import sys
 from typing import Any, Dict, Sequence, Tuple, Union, Container, NewType, Type, TypeVar
 import typing
 import typing_extensions
-import mypy_extensions
+
+try:
+    import mypy_extensions
+except ImportError:
+    mypy_extensions = None
 
 T = TypeVar("T")
 
@@ -105,6 +109,7 @@ if sys.version_info >= (3, 10):
     def is_newtype(obj: object) -> bool:
         return isinstance(obj, NewType)
 
+
 else:
 
     def is_newtype(obj: object) -> bool:
@@ -138,6 +143,8 @@ def _fill_typing_name_cache(name: str) -> Tuple[Tuple[Any, ...], Tuple[str, ...]
         objs = []
         names = []
         for mod in (typing, typing_extensions, mypy_extensions):
+            if mod is None:
+                continue
             try:
                 objs.append(getattr(mod, name))
                 names.append(f"{mod}.{name}")